mobile_convert/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mobile-convert"
version = "0.1.0"
description = "CLI-first retraining/export pipeline for mobile rice quality model."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "numpy>=1.26",
  "pandas>=2.2",
  "torch>=2.2",
  "torchvision>=0.17",
  "timm>=1.0.24",
  "albumentations>=2.0.8",
  "scikit-learn>=1.4",
  "pillow>=10",
  "tqdm>=4.66",
  "PyYAML>=6.0.1",
  "onnx>=1.16",
  "onnxruntime>=1.18",
  "onnxconverter-common>=1.14.0"
]

[project.scripts]
mobile-convert = "mobile_convert.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["mobile_convert*"]
